requirements.txt
fastapi
uvicorn[standard]
mcp
asyncpg
httpx
pydantic
cryptography
PyJWT
python-dotenv
sqlalchemy
psycopg2-binary